server/fix: 페이지네이션 알고리즘 수정

+++ packages/server/src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { HOST, PORT, PROTOCOL } from 'src/config/env';
 
 import { getRandomNickname } from '@woowa-babble/random-nickname';
@@ -14,6 +14,40 @@ export class UsersService {
     @InjectRepository(UserModel)
     private readonly usersRepository: Repository<UserModel>,
   ) {}
+
+  public async paginateUsers(paginateUserDto: PaginateUserDto) {
+    // 1. 커서 기반 페이지네이션
+    const users = await this.usersRepository.find({
+      where: { id: MoreThan(Number(paginateUserDto.where__id_more_than ?? 0)) },
+      order: { createdAt: paginateUserDto.order__created_at ?? 'ASC' },
+      take: Number(paginateUserDto.take ?? 5),
+    });
+
+    // 2. 다음 페이지 URL 생성
+    const nextUrl = new URL(`${PROTOCOL}://${HOST}:${PORT}/users`);
+    const lastItem = users[users.length - 1] ?? null;
+
+    // where__id_more_than 을 제외한 쿼리 파라미터는 그대로 유지
+    for (const queryKey of Object.keys(paginateUserDto)) {
+      if (queryKey === 'where__id_more_than') nextUrl.searchParams.set(queryKey, lastItem.id.toString());
+      else nextUrl.searchParams.append(queryKey, paginateUserDto[queryKey]);
+    }
+
+    // where__id_more_than이 존재하지 않으면 추가
+    if (!Object.keys(paginateUserDto).includes('where__id_more_than')) {
+      nextUrl.searchParams.set('where__id_more_than', lastItem.id.toString());
+    }
+
+    return {
+      data: users,
+      pageInfo: {
+        size: users.length,
+        lastCursor: lastItem?.id ?? null,
+        // DB 에서 가져온 데이터 개수보다 페이지네이션으로 요청한 개수가 더 적으면 다음 페이지 없음
+        nextUrl: users.length < (paginateUserDto.take || 5) ? null : nextUrl.toString(),
+      },
+    };
+  }
 
   public async createUser(createUserDto: CreateUserDto) {
     const newUser = this.usersRepository.create(createUserDto);
@@ -31,42 +65,9 @@ export class UsersService {
     return { message: `Successfully created ${size} dummy users` };
   }
 
-  public async paginateUsers(paginateUserDto: PaginateUserDto) {
-    // 커서 기반 페이지네이션
-    const users = await this.usersRepository.find({
-      where: {
-        id: MoreThan(Number(paginateUserDto.where__id_more_than ?? 0)),
-      },
-      order: {
-        createdAt: paginateUserDto.order__created_at,
-      },
-      take: Number(paginateUserDto.take ?? 5),
-    });
-
-    // 다음 페이지 URL 생성 (where__id_more_than 을 제외한 쿼리 파라미터는 그대로 유지)
-    const nextUrl = new URL(`${PROTOCOL}://${HOST}:${PORT}/users`);
-    const lastItem = users[users.length - 1] ?? null;
-
-    for (const queryKey of Object.keys(paginateUserDto)) {
-      if (queryKey === 'where__id_more_than')
-        nextUrl.searchParams.append(queryKey, lastItem.id.toString());
-      nextUrl.searchParams.append(queryKey, paginateUserDto[queryKey]);
-    }
-
-    return {
-      users,
-      pageInfo: {
-        size: users.length,
-        lastCursor: lastItem?.id ?? null,
-        nextUrl:
-          users.length < (paginateUserDto.take || 5)
-            ? null
-            : nextUrl.toString(),
-      },
-    };
-  }
-
   public async deleteUser(id: number) {
+    const user = await this.usersRepository.findOne({ where: { id } });
+    if (!user) throw new NotFoundException(`${id} 에 대한 유저가 존재하지 않습니다`);
     return this.usersRepository.delete(id);
   }
 }
